--- server/athenian/api/models/__init__.py
@@ -1,14 +1,0 @@
-# coding: utf-8
-
-# import models into model package
-from athenian.api.models.calculated_metric import CalculatedMetric
-from athenian.api.models.calculated_metric_values import CalculatedMetricValues
-from athenian.api.models.calculated_metrics import CalculatedMetrics
-from athenian.api.models.developer_set import DeveloperSet
-from athenian.api.models.for_set import ForSet
-from athenian.api.models.granularity import Granularity
-from athenian.api.models.invalid_request_error import InvalidRequestError
-from athenian.api.models.metric_id import MetricID
-from athenian.api.models.metrics_request import MetricsRequest
-from athenian.api.models.no_source_data_error import NoSourceDataError
-from athenian.api.models.repository_set import RepositorySet
