+++ server/athenian/api/models/web/__init__.py
@@ -1,0 +1,13 @@
+from athenian.api.models.web.calculated_metric import CalculatedMetric
+from athenian.api.models.web.calculated_metric_values import CalculatedMetricValues
+from athenian.api.models.web.calculated_metrics import CalculatedMetrics
+from athenian.api.models.web.created_identifier import CreatedIdentifier
+from athenian.api.models.web.developer_set import DeveloperSet
+from athenian.api.models.web.for_set import ForSet
+from athenian.api.models.web.generic_error import ForbiddenError, GenericError, NotFoundError
+from athenian.api.models.web.granularity import Granularity
+from athenian.api.models.web.invalid_request_error import InvalidRequestError
+from athenian.api.models.web.metric_id import MetricID
+from athenian.api.models.web.metrics_request import MetricsRequest
+from athenian.api.models.web.no_source_data_error import NoSourceDataError
+from athenian.api.models.web.repository_set import RepositorySet
