+++ server/athenian/api/models/web/created_identifier.py
@@ -3,7 +3,7 @@
 from typing import Optional
 
 from athenian.api import serialization
-from athenian.api.models.base_model_ import Model
+from athenian.api.models.web.base_model_ import Model
 
 
 class CreatedIdentifier(Model):
