+++ server/athenian/api/models/web/developer_set.py
@@ -1,23 +1,23 @@
 # coding: utf-8
 
 from athenian.api import serialization
-from athenian.api.models.base_model_ import Model
+from athenian.api.models.web.base_model_ import Model
 
 
-class RepositorySet(Model):
+class DeveloperSet(Model):
     """This class is auto generated by OpenAPI Generator (https://openapi-generator.tech)."""
 
     def __init__(self):
-        """RepositorySet - a model defined in OpenAPI"""
+        """DeveloperSet - a model defined in OpenAPI"""
         self.openapi_types = {}
 
         self.attribute_map = {}
 
     @classmethod
-    def from_dict(cls, dikt: dict) -> "RepositorySet":
+    def from_dict(cls, dikt: dict) -> "DeveloperSet":
         """Returns the dict as a model
 
         :param dikt: A dict.
-        :return: The RepositorySet of this RepositorySet.
+        :return: The DeveloperSet of this DeveloperSet.
         """
         return serialization.deserialize_model(dikt, cls)
