+++ server/athenian/api/models/web/granularity.py
@@ -6,7 +6,7 @@ from typing import List
 from dateutil.rrule import DAILY, MONTHLY, rrule, WEEKLY
 
 from athenian.api import serialization
-from athenian.api.models.base_model_ import Model
+from athenian.api.models.web.base_model_ import Model
 
 
 class Granularity(Model):
