+++ server/athenian/api/models/web/repository_set.py
@@ -1,23 +1,23 @@
 # coding: utf-8
 
 from athenian.api import serialization
-from athenian.api.models.base_model_ import Model
+from athenian.api.models.web.base_model_ import Model
 
 
-class DeveloperSet(Model):
+class RepositorySet(Model):
     """This class is auto generated by OpenAPI Generator (https://openapi-generator.tech)."""
 
     def __init__(self):
-        """DeveloperSet - a model defined in OpenAPI"""
+        """RepositorySet - a model defined in OpenAPI"""
         self.openapi_types = {}
 
         self.attribute_map = {}
 
     @classmethod
-    def from_dict(cls, dikt: dict) -> "DeveloperSet":
+    def from_dict(cls, dikt: dict) -> "RepositorySet":
         """Returns the dict as a model
 
         :param dikt: A dict.
-        :return: The DeveloperSet of this DeveloperSet.
+        :return: The RepositorySet of this RepositorySet.
         """
         return serialization.deserialize_model(dikt, cls)
